# springboot_openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
paths:
...
  /customers/{username}/wallet/recharge:
    post:
      tags:
      - customer-controller
      operationId: rechargeWallet
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      - name: euro
        in: query
        required: true
        schema:
          type: number
          format: float
      responses:
        "200":
          description: OK
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/feedback:
    post:
      tags:
      - customer-controller
      operationId: newFeedback
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      - name: text
        in: query
        required: true
        schema:
          type: string
      - name: vehicleId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /customers/{username}:
    get:
      tags:
      - customer-controller
      operationId: getCustomer
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DAOCustomerOnlyNoPassword"
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/transactions/last:
    get:
      tags:
      - customer-controller
      operationId: getLastTransactionByUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DAOTransactionOnly"
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/suspansions:
    get:
      tags:
      - customer-controller
      operationId: getSuspensionByUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOActiveSuspensionOnly"
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/races:
    get:
      tags:
      - customer-controller
      operationId: getRacesByUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAORaceOnly"
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/payments:
    get:
      tags:
      - customer-controller
      operationId: getPaymentsByCustomer
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOTransactionOnly"
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/greenPoints:
    get:
      tags:
      - customer-controller
      operationId: getGreenPointsByCustomer
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /customers/{username}/credit:
    get:
      tags:
      - customer-controller
      operationId: getCreditByCustomer
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
                format: float
        "403":
          description: Accesso negato per la risorsa richiesta in quanto non posseduta
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
components:
  schemas:
...
    DAOCustomerOnlyNoPassword:
      required:
      - cf
      - email
      - gender
      - name
      - surname
      - username
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        name:
          type: string
        surname:
          type: string
        gender:
          type: string
        cf:
          type: string
    DAOTransactionOnly:
      required:
      - id
      - time
      - value
      type: object
      properties:
        id:
          type: integer
          format: int32
        value:
          type: number
          format: float
        time:
          type: string
          format: date-time
    DAOActiveSuspensionOnly:
      required:
      - customerUsername
      - start
      - value
      type: object
      properties:
        customerUsername:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        value:
          type: number
          format: float
    DAORaceOnly:
      required:
      - customer
      - id
      - minDuration
      type: object
      properties:
        id:
          $ref: "#/components/schemas/RaceId"
        customer:
          type: string
        minDuration:
          type: integer
          format: int32
    RaceId:
      required:
      - idCustomer
      - idVehicle
      - start
      type: object
      properties:
        start:
          type: string
          format: date-time
        idCustomer:
          maxLength: 50
          minLength: 0
          type: string
        idVehicle:
          type: integer
          format: int32